02_complete_joystick/requirements.txt
pyserial==3.5
pyvjoy==1.0.1
